# handler: improve memory allocation and speed