@connor4312 2015-08-07
I think a ~~better~~ faster way would be to write a function that checks the validity of the string. The prefix is the only variant, and the information that's trying to be extracted would not be difficult to get at programmatically.