@fatih 2015-08-07
Yep, I just didn't want to mess up the logic for the first iteration. Imho, it can be improved probably without using a regexp at all.