@igm 2015-08-09
Thanks for this improvement @fatih, this PR is worth merging. As this change does not break API it looks safe to me to merge this PR in both master and v2 branches. But to be sure, do you or anyone see any reasons why this PR could not be included in v2? Also additional improvements as mentioned above in discussion are more that welcome.