@fatih 2015-08-10
Hi @igm. I think I made the PR directly to `v2`. Is it wrong? As you said, I think it's safe to apply it to all existing branches as there is no API change involved.